refactor(decoder): make NeuronForCausalLM a factory class
This is to prepare the support of multiple backends for this class.
